+++ cli/tools/registry/unfurl.rs
@@ -12,9 +12,10 @@ use deno_graph::DynamicTemplatePart;
 use deno_graph::ParserModuleAnalyzer;
 use deno_graph::TypeScriptReference;
 use deno_package_json::PackageJsonDepValue;
+use deno_resolver::sloppy_imports::SloppyImportsResolutionMode;
 use deno_runtime::deno_node::is_builtin_node_module;
 
-use crate::resolver::SloppyImportsResolver;
+use crate::resolver::CliSloppyImportsResolver;
 
 #[derive(Debug, Clone)]
 pub enum SpecifierUnfurlerDiagnostic {
@@ -42,14 +43,14 @@ impl SpecifierUnfurlerDiagnostic {
 }
 
 pub struct SpecifierUnfurler {
-  sloppy_imports_resolver: Option<SloppyImportsResolver>,
+  sloppy_imports_resolver: Option<CliSloppyImportsResolver>,
   workspace_resolver: WorkspaceResolver,
   bare_node_builtins: bool,
 }
 
 impl SpecifierUnfurler {
   pub fn new(
-    sloppy_imports_resolver: Option<SloppyImportsResolver>,
+    sloppy_imports_resolver: Option<CliSloppyImportsResolver>,
     workspace_resolver: WorkspaceResolver,
     bare_node_builtins: bool,
   ) -> Self {
@@ -179,7 +180,7 @@ impl SpecifierUnfurler {
     let resolved =
       if let Some(sloppy_imports_resolver) = &self.sloppy_imports_resolver {
         sloppy_imports_resolver
-          .resolve(&resolved, deno_graph::source::ResolutionMode::Execution)
+          .resolve(&resolved, SloppyImportsResolutionMode::Execution)
           .map(|res| res.into_specifier())
           .unwrap_or(resolved)
       } else {
@@ -388,6 +389,8 @@ fn to_range(
 mod tests {
   use std::sync::Arc;
 
+  use crate::resolver::SloppyImportsCachedFs;
+
   use super::*;
   use deno_ast::MediaType;
   use deno_ast::ModuleSpecifier;
@@ -455,7 +458,9 @@ mod tests {
     );
     let fs = Arc::new(RealFs);
     let unfurler = SpecifierUnfurler::new(
-      Some(SloppyImportsResolver::new(fs)),
+      Some(CliSloppyImportsResolver::new(SloppyImportsCachedFs::new(
+        fs,
+      ))),
       workspace_resolver,
       true,
     );
