+++ resolvers/deno/README.md
@@ -1,0 +1,3 @@
+# deno_resolver
+
+Deno resolution algorithm.
