+++ resolvers/node/analyze.rs
