+++ resolvers/node/clippy.toml
