+++ resolvers/node/package_json.rs
